# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Backend untuk aplikasi Creacipe - Platform berbagi resep masakan.",
        "title": "Creacipe API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/comments/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus komentar milik sendiri",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Komentar"
                ],
                "summary": "Hapus komentar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID komentar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/editor/categories": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat kategori tag baru (admin/editor only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kategori"
                ],
                "summary": "Buat kategori baru",
                "parameters": [
                    {
                        "description": "Data kategori",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/editor/categories/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengubah kategori yang ada (admin/editor only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kategori"
                ],
                "summary": "Update kategori",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID kategori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data kategori",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCategoryInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus kategori (admin/editor only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kategori"
                ],
                "summary": "Hapus kategori",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID kategori",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/menus/{id}/comments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua komentar untuk suatu resep (termasuk replies nested)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Komentar"
                ],
                "summary": "Ambil komentar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID resep",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat komentar di resep. Bisa reply komentar lain dengan parent_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Komentar"
                ],
                "summary": "Tambah komentar",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID resep",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data komentar",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "comment_text": {
                                    "type": "string"
                                },
                                "parent_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateCategoryInput": {
            "type": "object",
            "required": [
                "category_name"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateCategoryInput": {
            "type": "object",
            "required": [
                "category_name"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}